+++ Source/wall.cpp
@@ -3,7 +3,7 @@
 
 void Wall::Render(Resource& texture) const noexcept {
 	DrawTextureEx(texture.texture, {position.x - 100.0f, position.y	- 100.0f}, 0, 0.3f, WHITE);
-	DrawText(TextFormat("%i", health), position.x - 21, position.y + 10, 40, RED);
+	DrawText(TextFormat("%i", health), static_cast<int> (position.x) - 21, static_cast<int>(position.y) + 10, 40, RED);
 }
 
 void Wall::Update() noexcept {
